requirements.txt
Flask
requests
google-generativeai
langchain
langchain-google-genai
line-bot-sdk
